moved call graph examples
